moved ServiceIdentifier into own domain
