+++ src/main/java/com/github/softwareforstudents/schoolinternalcloud/common/web/shared/DiscoveryRoute.java
@@ -5,6 +5,7 @@ import java.io.OutputStream;
 import java.nio.charset.Charset;
 import java.util.List;
 
+import com.github.softwareforstudents.schoolinternalcloud.common.service.ServiceIdentifier;
 import com.github.softwareforstudents.schoolinternalcloud.common.web.HttpStatusCodes;
 import com.github.softwareforstudents.schoolinternalcloud.common.web.RouteHandler;
 import com.sun.net.httpserver.HttpExchange;
